+++ src/onegov.gazette/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Allows publishers to manage organizations and categories.
+  [msom]
+
 1.10.3 (2017-11-16)
 ~~~~~~~~~~~~~~~~~~~